yaya-ai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "yaya-ai"
version = "0.1.0"
description = "Yaya AI — Multimodal Foundation Model trained from scratch"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}

dependencies = [
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "numpy>=1.26.0",
    "sentencepiece>=0.2.0",
    "tokenizers>=0.19.0",
    "safetensors>=0.4.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
train = [
    "deepspeed>=0.14.0",
    "wandb>=0.16.0",
    "datasets>=2.18.0",
    "accelerate>=0.28.0",
    "flash-attn>=2.5.0",
    "triton>=2.2.0",
]
eval = [
    "lm-eval>=0.4.0",
    "scikit-learn>=1.4.0",
    "nltk>=3.8.0",
]
serve = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
